Ignore unnecessary files

+++ test_coverage.py
@@ -8,11 +8,11 @@ env["RUSTFLAGS"] = "-Cinstrument-coverage"
 env["LLVM_PROFILE_FILE"] = "tests-%p-%m.profraw"
 
 # clean the project
-# retcode = subprocess.call(["cargo", "clean"], env=env)
-#
-# if retcode != 0:
-#     print("Failed to clean the project! Exiting!")
-#     exit(1)
+retcode = subprocess.call(["cargo", "clean"], env=env)
+
+if retcode != 0:
+    print("Failed to clean the project! Exiting!")
+    exit(1)
 
 # remove coverage folder
 shutil.rmtree("coverage")
@@ -24,7 +24,7 @@ if retcode != 0:
     print("Tests failed!")
 
 # create coverage data
-retcode = subprocess.call(["grcov", ".", "--binary-path", "./target/debug/", "-s", ".", "-t", "html", "--branch", "--ignore-not-existing", "-o", "./coverage/"])
+retcode = subprocess.call(["grcov", ".", "--binary-path", "./target/debug/", "-s", ".", "-t", "html", "--ignore", "*lab.rs", "--ignore", "*main.rs", "--ignore-not-existing", "-o", "./coverage/"])
 
 if retcode != 0:
     print("Failed to generate coverage data! Exiting!")
